Fix deprecation warnings in Python 3.9
See:
- https://docs.python.org/3/whatsnew/3.9.html#you-should-check-for-deprecationwarning-in-your-code
- https://bugs.python.org/issue36048

+++ extra_foam/gui/ctrl_widgets/geometry_ctrl_widget.py
@@ -161,8 +161,8 @@ class GeometryCtrlWidget(_AbstractCtrlWidget):
             header.setSectionResizeMode(i, QHeaderView.Stretch)
 
         header_height = table.horizontalHeader().height()
-        table.setMinimumHeight(header_height * (n_row + 1.5))
-        table.setMaximumHeight(header_height * (n_row + 2.0))
+        table.setMinimumHeight(int(header_height * (n_row + 1.5)))
+        table.setMaximumHeight(int(header_height * (n_row + 2.0)))
 
     def _setGeometryFile(self):
         filepath = QFileDialog.getOpenFileName()[0]
