+++ extra_foam/special_suite/xas_tim_xmcd_w.py
@@ -229,7 +229,7 @@ class XasTimXmcdWindow(_SpecialAnalysisBase):
 
         cw = self.centralWidget()
         cw.addWidget(right_panel)
-        cw.setSizes([self._TOTAL_W / 4, 3 * self._TOTAL_W / 4])
+        cw.setSizes([int(self._TOTAL_W / 4), int(3 * self._TOTAL_W / 4)])
 
         self.resize(self._TOTAL_W, self._TOTAL_H)
